requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
motor>=3.6.0
pymongo>=4.9.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic-settings>=2.6.0
httpx>=0.27.0
